#40 プラン内容保存処理
