+++ backend/routes/api.php
@@ -5,7 +5,6 @@ use App\Http\Controllers\Auth\RegisterController;
 use App\Http\Controllers\PlanController;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Route;
-use App\Http\Controllers\PlanDayController;
 use App\Http\Controllers\PlanDetailController;
 
 // 認証が必要なAPIルート
@@ -15,8 +14,7 @@ Route::middleware('auth:sanctum')->group(function () {
     });
     Route::post('/logout', [LoginController::class, 'logout']);
     Route::apiResource('plans', PlanController::class);
-    Route::apiResource('plans.days', PlanDayController::class);
-    Route::apiResource('plans.days.details', PlanDetailController::class);
+    Route::apiResource('plans/details', PlanDetailController::class);
 });
 
 // 認証が不要なAPIルート
